machine-learning/Docker_API/requirements.txt
tensorflow==2.18.0
scikit-learn
numpy
pandas
opencv-python
pillow
fastapi
uvicorn
python-multipart